refactor: restructure frontend architecture
Separated larger components into smaller ones, better distributed logic, and implemented a service for future use with useContext.
