+++ src/components/spinner/spinner.jsx
@@ -1,0 +1,13 @@
+import Spinner from 'react-bootstrap/Spinner';
+
+function LoadingSpinner () {
+  return (
+    <div className="spinner-container">
+      <Spinner animation="border" role="status">
+        <span className="visually-hidden">Loading...</span>
+      </Spinner>
+    </div>
+  );
+}
+
+export default LoadingSpinner;
