+++ src/services/questions/getQuestions.js
@@ -1,0 +1,12 @@
+export const getQuestions = async () => {
+  try {
+    const response = await fetch('http://localhost:3000/questions');
+    if (!response.ok) {
+      throw new Error('Network response was not ok');
+    }
+    const data = await response.json();
+    return data;
+  } catch (err) {
+    throw new Error('Error al cargar las preguntas iniciales: ' + err.message);
+  }
+};
